Some UI changes

+++ app/src/main/java/com/mi/simple_alarm_clock_app/ui/fragments/List/AlarmListFragment.java
@@ -1,19 +1,13 @@
 package com.mi.simple_alarm_clock_app.ui.fragments;
 
-import android.Manifest;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.opengl.Visibility;
-import android.os.Build;
 import android.os.Bundle;
 
-import androidx.activity.result.ActivityResultLauncher;
-import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.view.MenuProvider;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
@@ -22,7 +16,6 @@ import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.LinearLayoutManager;
 
 import android.provider.Settings;
-import android.util.Log;
 import android.view.ActionMode;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -30,28 +23,16 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.mi.simple_alarm_clock_app.R;
 import com.mi.simple_alarm_clock_app.Tools;
 import com.mi.simple_alarm_clock_app.alarmclock.AlarmClockManager;
-import com.mi.simple_alarm_clock_app.alarmclock.TimeUtils;
-import com.mi.simple_alarm_clock_app.database.DatabaseManager;
 import com.mi.simple_alarm_clock_app.databinding.FragmentAlarmListBinding;
 import com.mi.simple_alarm_clock_app.model.Alarm;
 import com.mi.simple_alarm_clock_app.ui.fragments.List.AlarmListFragmentViewModelFactory;
 import com.mi.simple_alarm_clock_app.ui.fragments.List.ListAdapter;
 import com.mi.simple_alarm_clock_app.ui.fragments.List.ListViewModel;
-
-import java.util.ArrayList;
-
-import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers;
-import io.reactivex.rxjava3.core.Completable;
-import io.reactivex.rxjava3.core.Observable;
-import io.reactivex.rxjava3.core.Single;
-import io.reactivex.rxjava3.disposables.Disposable;
-import io.reactivex.rxjava3.schedulers.Schedulers;
 
 public class AlarmListFragment extends Fragment implements MenuProvider, ListAdapter.ListAdapterListener {
 
@@ -129,12 +110,12 @@ public class AlarmListFragment extends Fragment implements MenuProvider, ListAda
     private void setNextAlarmTimeInUI() {
         try {
             long timeInMillis = new AlarmClockManager(context).getNextAlarmTimeInMillis();
-            binding.nextAlarmInscription.setText(getString(R.string.next_alarm_inscription));
+            binding.nextAlarmInscription.setText(getString(R.string.next_alarm_inscription_for_list));
             binding.nextAlarmTime.setText(Tools.getFormattedTimeTittleFromMillis(timeInMillis));
             binding.nextAlarmDate.setText(Tools.getDateTittleFromMillis(timeInMillis));
         } catch (NullPointerException e) {
             binding.nextAlarmInscription.setText("");
-            binding.nextAlarmTime.setText("0:0");
+            binding.nextAlarmTime.setText(getString(R.string.time_zero_zero));
             binding.nextAlarmDate.setText("Нет установленных будильников");
         }
     }
@@ -174,14 +155,42 @@ public class AlarmListFragment extends Fragment implements MenuProvider, ListAda
     @Override
     public boolean onMenuItemSelected(@NonNull MenuItem menuItem) {
         int itemId = menuItem.getItemId();
-        if (itemId == R.id.miSettings) {
-            navigate(R.id.action_firstFragment_to_alarmEditFragment);
+        if (itemId == R.id.miAdd) {
+            if (!Settings.canDrawOverlays(requireActivity())) {
+                MaterialAlertDialogBuilder dialogBuilder = new MaterialAlertDialogBuilder(context)
+                        .setTitle(getString(R.string.can_draw_overlays_alert_tittle))
+                        .setMessage(getString(R.string.can_draw_overlays_alert_message))
+                        .setCancelable(false)
+                        .setPositiveButton(getString(R.string.ok), ((dialog, which) -> {
+                            Intent intent = new Intent(Settings.ACTION_MANAGE_OVERLAY_PERMISSION, Uri.parse(
+                                    "package:" + context.getPackageName()
+                            ));
+                            startActivityForResult(intent, 9);
+                        }));
+                AlertDialog alertDialog = dialogBuilder.create();
+                alertDialog.show();
+            } else {
+                navigate(R.id.action_firstFragment_to_alarmEditFragment);
+            }
         }
         return true;
     }
 
     private void checkActivityExtras() {
-        Tools.checkActivityExtras(context, requireActivity().getIntent());
+        Intent intent = requireActivity().getIntent();
+        boolean extrasExist = Tools.checkActivityExtras(intent);
+        if (extrasExist) {
+            MaterialAlertDialogBuilder alertDialogBuilder = new MaterialAlertDialogBuilder(context)
+                    .setTitle(context.getString(R.string.attention))
+                    .setMessage(context.getString(R.string.time_zone_changed))
+                    .setPositiveButton(context.getString(R.string.ok), (dialog, which) -> {
+                        // nothing
+                    });
+            AlertDialog alertDialog = alertDialogBuilder.create();
+            alertDialog.show();
+
+            intent.removeExtra("time_zone_changed");
+        }
     }
 
     private void navigate(int destination) {
@@ -211,7 +220,16 @@ public class AlarmListFragment extends Fragment implements MenuProvider, ListAda
                 int itemId = menuItem.getItemId();
 
                 if (itemId == R.id.miDelete) {
-                    viewModel.deleteSelectedAlarmsInActionMode();
+                    MaterialAlertDialogBuilder dialogBuilder = new MaterialAlertDialogBuilder(context)
+                            .setTitle(getString(R.string.confirm))
+                                    .setPositiveButton(getString(R.string.delete), ((dialog, which) -> {
+                                        viewModel.deleteSelectedAlarmsInActionMode();
+                                    }))
+                            .setNegativeButton(getString(R.string.cansel), ((dialog, which) -> {
+                                // nothing
+                            }));
+                    AlertDialog alertDialog = dialogBuilder.create();
+                    alertDialog.show();
                 }
 
                 return true;
